Move item comparison into a module
