+++ src/scss/_style.scss
@@ -1282,112 +1282,8 @@ g {
   }
 }
 
-dim-compare {
-  #loadout-drawer {
-    padding-left: 1em;
-    padding-right: 1em;
-  }
-}
-
 .loadout-item {
   position: relative;
-}
-
-.compare-bucket {
-  white-space: nowrap;
-  overflow-x: scroll;
-  display: flex;
-  margin-left: 100px;
-  flex-direction: row;
-  justify-content: flex-start;
-  .fixed-left {
-    border-left: none;
-    position: fixed;
-    left: 1em;
-    width: 100px;
-  }
-  .compare-stat-label {
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  .talent-grid {
-    padding: 4px 2px 0 2px;
-    box-sizing: border-box;
-    width: 150px;
-    height: auto;
-  }
-}
-
-.compare-options {
-  margin: 1em 0;
-}
-
-.compare-item {
-  border-left: 1px solid $lightgray;
-  width: 150px;
-
-  .item {
-    float: right;
-    padding: 0 4px 0 0;
-    margin-right: 4px;
-    z-index: -10;
-  }
-  > div {
-    padding-left: 4px;
-  }
-  .highlight {
-    background-color: rgba(128, 128, 128, .25);
-    cursor: pointer;
-  }
-  .item-name {
-    text-overflow: ellipsis;
-    overflow: hidden;
-    cursor: pointer;
-    margin-top: 2px;
-    margin-bottom: 2px;
-
-    .elemental {
-      vertical-align: bottom;
-      margin: 0;
-    }
-  }
-  .sorted {
-    color: $orange;
-  }
-  .close {
-    position: static;
-    float: right;
-    margin: 1px 4px;
-    padding-left: 0px;
-  }
-  .range {
-    font-size: .9em;
-  }
-  dim-item-tag {
-    height: 20px;
-    margin-left: 5px;
-    display: block;
-    select {
-      background: gray;
-      border: none;
-      color: white;
-      width: 120px;
-    }
-  }
-  .spacer {
-    height: 20px;
-    &.withTags {
-      height: 40px;
-    }
-  }
-}
-
-@keyframes pop {
-  to { transform: scale(1.5) }
-}
-.item-pop {
-  animation: .25s linear 2 alternate pop;
-  z-index: 1;
 }
 
 .loadout-content {
